Manager: split out toolkit API into sub-module
